+++ src/elements/listformat/listformat.spec.ts
@@ -52,6 +52,22 @@ describe('intl-listformat', () => {
     expect(getTextContent(page.element)).toBe(intlResult);
   });
 
+  it('has `value` property the same as its text content', async () => {
+    const page = await createTestPage<HTMLIntlListFormatElement>({
+      element: 'intl-listformat',
+      html: `
+        <intl-listformat locales="ar">
+          <intl-listitem>foo</intl-listitem>
+          <intl-listitem>bar</intl-listitem>
+          <intl-listitem>baz</intl-listitem>
+          <intl-listitem>qux</intl-listitem>
+        </intl-listformat>
+      `,
+    });
+
+    expect(page.element!.value).toBe(getTextContent(page.element));
+  });
+
   it('handles multiple locales and ignores invalid ones', async () => {
     const page = await createTestPage<HTMLIntlListFormatElement>({
       element: ['intl-listformat', 'intl-listitem'],
@@ -102,16 +118,26 @@ describe('intl-listformat', () => {
         textContent: 'world',
       }),
     );
+    el.querySelector('intl-listitem:nth-child(2)')!.remove();
 
     // TODO: look into why this triggers 2 update cycles.
     await el.updateComplete;
     await el.updateComplete;
 
-    expect(el.list).toEqual(['foo', 'bar', 'baz', 'qux', 'hello', 'world']);
-
+    expect(el.list).toEqual(['foo', 'baz', 'qux', 'hello', 'world']);
     // @ts-ignore
     intlResult = new Intl.ListFormat('de').format([
-      'foo', 'bar', 'baz', 'qux', 'hello', 'world']);
+        'foo', 'baz', 'qux', 'hello', 'world']);
+    expect(getTextContent(el)).toBe(intlResult);
+
+    el.querySelector('intl-listitem:last-child')!.textContent = 'moon';
+    await el.updateComplete;
+    await el.updateComplete;
+
+    expect(el.list).toEqual(['foo', 'baz', 'qux', 'hello', 'moon']);
+    // @ts-ignore
+    intlResult = new Intl.ListFormat('de').format([
+        'foo', 'baz', 'qux', 'hello', 'moon']);
     expect(getTextContent(el)).toBe(intlResult);
   });
 
@@ -159,5 +185,47 @@ describe('intl-listformat', () => {
     }).formatToParts(['foo', 'bar', 'baz', 'qux']);
 
     expect(el.formatToParts()).toEqual(intlResult);
+  });
+
+  it('trims list item text', async () => {
+    const page = await createTestPage<HTMLIntlListFormatElement>({
+      element: ['intl-listformat', 'intl-listitem'],
+      html: `
+        <intl-listformat locales="de">
+          <intl-listitem> foo </intl-listitem>
+          <intl-listitem>bar </intl-listitem>
+          <intl-listitem> baz   </intl-listitem>
+          <intl-listitem> qux </intl-listitem>
+        </intl-listformat>
+      `,
+    });
+    const el = page.element;
+
+    // @ts-ignore
+    let intlResult = new Intl.ListFormat('de').format([
+      'foo', 'bar', 'baz', 'qux']);
+
+    expect(getTextContent(el)).toBe(intlResult);
+  });
+
+  it('ignores empty list items', async () => {
+    const page = await createTestPage<HTMLIntlListFormatElement>({
+      element: ['intl-listformat', 'intl-listitem'],
+      html: `
+        <intl-listformat locales="de">
+          <intl-listitem>foo</intl-listitem>
+          <intl-listitem></intl-listitem>
+          <intl-listitem>baz</intl-listitem>
+          <intl-listitem></intl-listitem>
+        </intl-listformat>
+      `,
+    });
+    const el = page.element;
+
+    // @ts-ignore
+    const intlResult = new Intl.ListFormat('de').format(['foo', 'baz']);
+
+    expect(el.list).toEqual(['foo', 'baz']);
+    expect(getTextContent(el)).toBe(intlResult);
   });
 });